musiquepy_website/requirements.txt
flask==2.0.2
pytest==6.2.5
WTForms==3.0.0
autopep8==1.6.0
python-dotenv==0.19.2
email-validator==1.1.3
Flask-Session==0.4.0